refactor: 필터 기능 상위 컴포넌트로 이동
